+++ pytorch_lightning/core/decorators.py
@@ -17,7 +17,6 @@
 from functools import wraps
 from typing import Callable
 
-from pytorch_lightning.core.lightning import LightningModule
 from pytorch_lightning.utilities import rank_zero_warn
 
 
@@ -57,6 +56,7 @@ def auto_move_data(fn: Callable) -> Callable:
     """
     @wraps(fn)
     def auto_transfer_args(self, *args, **kwargs):
+        from pytorch_lightning.core.lightning import LightningModule
         if not isinstance(self, LightningModule):
             return fn(self, *args, **kwargs)
 
@@ -70,17 +70,15 @@ def auto_move_data(fn: Callable) -> Callable:
 def parameter_validation(fn: Callable) -> Callable:
     @wraps(fn)
     def inner_f(self, *args, **kwargs):
-        if not isinstance(self, LightningModule):
-            return fn(self, *args, **kwargs)
         pre_param_count = len(list(self.parameters()))
         module = fn(self, *args, **kwargs)
         self.on_post_move_to_device()
         post_param_count = len(list(self.parameters()))
 
         if not pre_param_count == post_param_count:
-            rank_zero_warn('The parameter count does not match after moving target device. '
+            rank_zero_warn('The model parameters do not match after moving to the target device. '
                            'If your model employs weight sharing on TPU,'
-                           'please tie your weights in the `on_post_move_to_device` model hook.')
+                           'please tie your weights using the `on_post_move_to_device` model hook.')
 
         return module